Breaking out policy to own plugin and Fixing policy bug service call check using URI and not resolved_name
